fix threshold_count max length error

+++ ind_initialize.py
@@ -5,4 +5,4 @@ def ind_initialize(max_size, N, step):
     ind = range(N, max_size-N, step)
     if ind[-1] < max_size - N - 1:
         ind = np.append(ind, np.array([max_size - N - 1]), axis=0)
-    return ind.astype(np.int)
+    return ind
